[revert] (session-authorization) remove sessionUtils
