+++ src/main/java/com/woojin/userdemo/user/SessionRefreshFilter.java
@@ -11,4 +11,5 @@ import org.springframework.web.filter.OncePerRequestFilter;
 @Component
 @RequiredArgsConstructor
 public class SessionRefreshFilter extends OncePerRequestFilter {
+    private final
 }
